evaluator/evaluator_test.go
package evaluator

import (
	"testing"
	"woc_lang/lexer"
	"woc_lang/object"
	"woc_lang/parser"
)

func TestEvalIntegerExpression(t *testing.T) {
	tests := []struct {
		input    string
		expected int64
	}{
		{"5;", 5},
		{"10;", 10},
	}

	for _, tt := range tests {
		evaluated := testEval(t, tt.input)
		testIntegerObject(t, evaluated, tt.expected)
	}
}

func testEval(t *testing.T, input string) object.Object {
	l := lexer.New(input)
	checkLexerErrors(t, l)
	p := parser.New(l)
	checkParserErrors(t, p)
	program := p.Program

	return Eval(program)
}

func testIntegerObject(t *testing.T, obj object.Object, expected int64) bool {
	result, ok := obj.(*object.Integer)
	if !ok {
		t.Errorf("对象不是 Integer 类型，实际获得: %T (%+v)", obj, obj)
		return false
	}

	if result.Value != expected {
		t.Errorf("Integer 对象值错误:\n实际获得: %d\n期望获得: %d", result.Value, expected)
		return false
	}

	return true
}

func checkParserErrors(t *testing.T, p *parser.Parser) {
	errMessages := p.Errors()
	if len(errMessages) == 0 {
		return
	}

	t.Errorf("语法分析存在错误")
	for _, msg := range errMessages {
		t.Errorf("语法法分析错误: %q", msg)
	}
	t.FailNow()
}

func checkLexerErrors(t *testing.T, l *lexer.Lexer) {
	errTokens := l.Errors()
	if len(errTokens) == 0 {
		return
	}

	t.Errorf("词法分析器存在 %d 个错误", len(errTokens))
	for _, errTok := range errTokens {
		t.Errorf("(%s) 词法分析错误: %q", errTok.Literal, errTok.Msg)
	}
	t.FailNow()
}
